+++ app/blueprints/address_search/__init__.py
@@ -1,0 +1,4 @@
+from .api import address_search_api_bp
+from .pages import address_search_pages_bp
+
+__all__ = ["address_search_api_bp", "address_search_pages_bp"]